Advance of linear history for update of osl/pub/stream.next to commit 0164819be080dc4e7d39961731d279677843101c

+++ hw/xilinx_axidma.c
@@ -393,7 +393,7 @@ xilinx_axidma_control_stream_push(StreamSlave *obj, unsigned char *buf,
 
     if (len != CONTROL_PAYLOAD_SIZE) {
         hw_error("AXI DMA requires %d byte control stream payload\n",
-                 CONTROL_PAYLOAD_SIZE);
+                 (int)CONTROL_PAYLOAD_SIZE);
     }
 
     memcpy(s->app, buf, len);
